HierarchyCode/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
